Report real class names; keep the leniency on the inbound side
Follow-up on review: reporting a de-"_C"'d class name misrepresents what
the class actually is. Class names now go out as the real
GetClass()->GetName() everywhere, Blueprint "_C" and all, and
GetSubClassWithName alone reconciles the two spellings on the way in.

- Drop UTempoCoreUtils::GetClassIdentifier/GetClassIdentifierName. The
  WorldControl service is back to reporting GetClass()->GetName()
  unchanged; the round trip works because the resolver is lenient, not
  because we renamed anything.
- Move GetSubClassWithName from TempoWorldUtils.h to a new
  TempoCore/Public/TempoClassUtils.h so every plugin can resolve a
  class name the same way (TempoSensors depends on TempoCore, not
  TempoWorld). TempoWorldUtils.h includes it, so existing users of that
  header are unaffected. AssetRegistry becomes a public TempoCore dep.
- SetActorTypeSemanticId resolves its actor_type through the shared
  resolver and keys the override map on Class->GetFName(), so either
  spelling is accepted, everything the labeler reports is a real class
  name, and an unresolvable name returns NOT_FOUND instead of quietly
  registering an override no actor will ever match. LabelActor's lookup
  is back to a plain FName hash.

StripBlueprintClassSuffix stays as the one piece the resolver needs, now
documented as inbound-only.

Note GetActorIdentifier is unchanged and does no string surgery: it is
GetActorNameOrLabel() with the label materialized so the value is stable.
The "_C" that disappears in editor builds is Unreal's own
GetDefaultActorLabel() doing RemoveFromEnd("_C"); a packaged build has no
labels and reports the real object name.
