+++ TempoSensors/Source/TempoSensors/Private/TempoActorLabeler.cpp
@@ -9,6 +9,7 @@
 
 #include "TempoSensorsSettings.h"
 
+#include "TempoClassUtils.h"
 #include "TempoCoreUtils.h"
 #include "DefaultActorClassifier.h"
 
@@ -141,7 +142,7 @@ void UTempoActorLabeler::HandleGetSemanticClasses(const TempoCore::Empty& Reques
 	{
 		if (const int32* SemanticId = SemanticIds.Find(LabelName))
 		{
-			SemanticIdToActorTypes.FindOrAdd(*SemanticId).Add(UTempoCoreUtils::GetClassIdentifierName(ActorClass));
+			SemanticIdToActorTypes.FindOrAdd(*SemanticId).Add(ActorClass->GetFName());
 		}
 	}
 
@@ -208,9 +209,6 @@ void UTempoActorLabeler::HandleGetSemanticClasses(const TempoCore::Empty& Reques
 
 void UTempoActorLabeler::HandleSetActorTypeSemanticId(const TempoSensors::SetActorTypeSemanticIdRequest& Request, const TResponseDelegate<TempoCore::Empty>& ResponseContinuation)
 {
-	// Accept the class name with or without the Blueprint "_C" suffix, and key the override on the
-	// same identifier GetAllActorLabels/GetSemanticClasses report back.
-	const FName ActorType = FName(*UTempoCoreUtils::StripBlueprintClassSuffix(FString(UTF8_TO_TCHAR(Request.actor_type().c_str()))));
 	const int32 SemanticId = Request.semantic_id();
 
 	// Validate range
@@ -221,6 +219,20 @@ void UTempoActorLabeler::HandleSetActorTypeSemanticId(const TempoSensors::SetAct
 			"semantic_id must be -1 (revert) or 0-255"));
 		return;
 	}
+
+	// Resolve to an actual class so that either spelling of a Blueprint class name is accepted, the
+	// override is keyed on the real class name we report elsewhere, and a name we can't place is
+	// rejected rather than silently registering an override no actor will ever match.
+	const FString ActorTypeName(UTF8_TO_TCHAR(Request.actor_type().c_str()));
+	const UClass* ActorClass = GetSubClassWithName<AActor>(ActorTypeName);
+	if (!ActorClass)
+	{
+		const FString ErrorMsg = FString::Printf(TEXT("No actor class with name '%s' found"), *ActorTypeName);
+		ResponseContinuation.ExecuteIfBound(TempoCore::Empty(),
+			grpc::Status(grpc::StatusCode::NOT_FOUND, std::string(TCHAR_TO_UTF8(*ErrorMsg))));
+		return;
+	}
+	const FName ActorType = ActorClass->GetFName();
 
 	// Store or clear override
 	if (SemanticId < 0)
@@ -235,7 +247,7 @@ void UTempoActorLabeler::HandleSetActorTypeSemanticId(const TempoSensors::SetAct
 	// Re-label all existing actors of this type
 	for (TActorIterator<AActor> ActorItr(GetWorld()); ActorItr; ++ActorItr)
 	{
-		if (UTempoCoreUtils::GetClassIdentifierName(ActorItr->GetClass()) == ActorType)
+		if (ActorItr->GetClass()->GetFName() == ActorType)
 		{
 			UnLabelActor(*ActorItr);
 			LabelActor(*ActorItr);
@@ -380,7 +392,7 @@ void UTempoActorLabeler::HandleGetAllActorLabels(const TempoCore::Empty& Request
 
 		auto* ActorInfo = Response.add_actors();
 		ActorInfo->set_actor_name(TCHAR_TO_UTF8(*UTempoCoreUtils::GetActorIdentifier(Actor)));
-		ActorInfo->set_actor_type(TCHAR_TO_UTF8(*UTempoCoreUtils::GetClassIdentifier(Actor->GetClass())));
+		ActorInfo->set_actor_type(TCHAR_TO_UTF8(*Actor->GetClass()->GetName()));
 		ActorInfo->set_semantic_id(IdPair.SemanticId);
 		ActorInfo->set_instance_id(IdPair.InstanceId);
 	}
@@ -578,10 +590,7 @@ void UTempoActorLabeler::LabelActor(AActor* Actor)
 		return;
 	}
 
-	// The name this Actor's class goes by over the API. Also the key of both maps below, so that a
-	// SetActorTypeSemanticId override registered under either spelling of a Blueprint class name
-	// applies here, and GetLabeledActorTypes reports the same spelling back.
-	const FName ActorTypeName = UTempoCoreUtils::GetClassIdentifierName(Actor->GetClass());
+	const FName ActorTypeName = Actor->GetClass()->GetFName();
 
 	// Check for type-level override first (before DataTable lookup)
 	if (const int32* TypeOverride = ActorTypeSemanticIdOverrides.Find(ActorTypeName))
