+++ TempoSensors/Source/TempoSensors/Public/Labels.proto
@@ -16,7 +16,7 @@ message InstanceToSemanticIdMap {
 message ActorLabelInfo {
   // Actor name, matching the identifier the TempoWorld RPCs use for this actor.
   string actor_name = 1;
-  // Actor class name, without the Blueprint "_C" suffix.
+  // Actor class name, e.g. "CBP_QuantumCharacter_C" for a Blueprint class.
   string actor_type = 2;
   int32 semantic_id = 3;
   int32 instance_id = 4;  // For Instance mode correlation
@@ -27,15 +27,14 @@ message GetAllActorLabelsResponse {
 }
 
 message GetLabeledActorTypesResponse {
-  // Actor class names that currently have active instance IDs in the scene,
-  // without the Blueprint "_C" suffix.
+  // Actor class names that currently have active instance IDs in the scene.
   repeated string types = 1;
 }
 
 message SemanticClassInfo {
   string name = 1;                        // e.g., "Pedestrian", "Car"
   int32 label_id = 2;                     // e.g., 12, 14 (the semantic ID)
-  repeated string actor_types = 3;        // Actor types assigned to this class (no Blueprint "_C" suffix)
+  repeated string actor_types = 3;        // Actor types assigned to this class
   repeated string static_mesh_types = 4;  // Static mesh paths assigned to this class
 }
 
@@ -44,8 +43,8 @@ message GetSemanticClassesResponse {
 }
 
 message SetActorTypeSemanticIdRequest {
-  // Class name, with or without the generated "_C" suffix for a Blueprint class.
-  string actor_type = 1;    // e.g., "CBP_QuantumCharacter"
+  // Class name. A Blueprint class may be named with or without its generated "_C" suffix.
+  string actor_type = 1;    // e.g., "CBP_QuantumCharacter_C" (class name)
   int32 semantic_id = 2;    // 0-255 to set, -1 to revert to DataTable default
 }
