+++ TempoWorld/Source/TempoWorld/Public/WorldControl.proto
@@ -6,8 +6,8 @@ import "TempoCore/Empty.proto";
 import "TempoCore/Geometry.proto";
 
 message SpawnActorRequest {
-  // Class name of the actor to spawn. Blueprint classes may be named with or without the
-  // generated "_C" suffix ("BP_Foo" or "BP_Foo_C"); an exact match is preferred.
+  // Class name of the actor to spawn. A Blueprint class may be named either by its generated
+  // class name ("BP_Foo_C") or by the asset name it reads as in the editor ("BP_Foo").
   string actor_type = 1;
   // If true, defer finishing the spawn until FinishSpawningActor is called.
   bool deferred = 2;
@@ -40,7 +40,7 @@ message DestroyActorRequest {
 }
 
 message AddComponentRequest {
-  // Class name of the component to add. Accepts the Blueprint "_C" suffix or not, as SpawnActor does.
+  // Class name of the component to add. Named the same way as SpawnActor's actor_type.
   string component_type = 1;
   // Name of the owning actor.
   string actor = 2;
@@ -105,7 +105,7 @@ message DeactivateComponentRequest {
 message ActorDescriptor {
   // Actor name.
   string name = 1;
-  // Actor class name, without the Blueprint "_C" suffix. Can be passed straight back to SpawnActor.
+  // Actor class name, e.g. "BP_Foo_C" for a Blueprint class. Can be passed straight back to SpawnActor.
   string actor_type = 2;
 }
 
@@ -118,7 +118,7 @@ message ComponentDescriptor {
   string actor = 1;
   // Component name.
   string name = 2;
-  // Component class name, without the Blueprint "_C" suffix. Can be passed straight back to AddComponent.
+  // Component class name, e.g. "BP_Foo_C" for a Blueprint class. Can be passed straight back to AddComponent.
   string component_type = 3;
 }
 
@@ -284,7 +284,7 @@ message SetClassPropertyRequest {
   string actor = 1;
   string component = 2;
   string property = 3;
-  // Class path or name. Accepts the Blueprint "_C" suffix or not, as SpawnActor does.
+  // Class path or name. Named the same way as SpawnActor's actor_type.
   string value = 4;
 }
 
@@ -358,7 +358,7 @@ message SetClassArrayPropertyRequest {
   string actor = 1;
   string component = 2;
   string property = 3;
-  // Class paths or names. Accept the Blueprint "_C" suffix or not, as SpawnActor does.
+  // Class paths or names. Named the same way as SpawnActor's actor_type.
   repeated string values = 4;
 }
 
@@ -431,7 +431,7 @@ message SetClassSetPropertyRequest {
   string actor = 1;
   string component = 2;
   string property = 3;
-  // Class paths or names. Accept the Blueprint "_C" suffix or not, as SpawnActor does.
+  // Class paths or names. Named the same way as SpawnActor's actor_type.
   repeated string values = 4;
 }
